+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Change Log
 
+## v1.3
+
+- コマンドラインオプション追加: `--config-file` `--config-section` `--create-config-file`
+- コマンドラインオプション変更: `--generate-samplefile` -> `--create-sample-datafile`
+
 ## v1.2
 
-- デプロイ枚数オプションの n, n_example をnullableに変更。nullのとき、全画像を出力。
+- 設定値 n, n_example をnullableに変更。nullのとき、全画像を出力。
 - コマンドラインオプション変更: `--makesample <N>` -> `--generate-samplefile`
 - コマンドラインオプション追加: `--deploy-result`
 - 設定値追加: figsize, backup
@@ -12,4 +17,4 @@
 
 - 自作カラーマップをサポート。
 - 設定値 vmax の初期値を変更。
-- 設定値 vmin, vmax の空欄設定(自動判定)をサポート。
+- 設定値 vmin, vmax をnullableに変更。nullのとき、自動判定。